requirements.txt
numpy
pandas
matplotlib
pillow
wordcloud
alpha=0.25